requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
SQLAlchemy>=2.0,<3.0
pydantic-settings>=2.6,<3.0
python-dotenv>=1.0,<2.0
google-auth>=2.36,<3.0
google-auth-oauthlib>=1.2,<2.0
google-api-python-client>=2.150,<3.0

